backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymysql==1.1.0
bcrypt==4.0.1
pyjwt==2.8.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==2.3.7
